Updates for edge cases of adding removing
If we have to add to a list of length 1, we manually increment size to
be 2. If we have to remove from a list of length 3 or 2 manually
decrement size by 1 in order to avoid getting stuck due to the int
arithmetic (3 * .25 and 2 * .25 both truncate to 0).

Also removed a use of double arithmetic in the remove method and
replaced with int calculations.